+++ app/config/global.h
@@ -70,6 +70,9 @@ namespace Config {
     int playlistRowHeight() const;
     void savePlaylistRowHeight(int arg);
 
+    bool showPlaylistHeaders() const;
+    void saveShowPlaylistHeaders(bool arg);
+
     bool waveformDisabled() const;
     void saveWaveformDisabled(bool arg);